# api/internal/repository/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package repository

import (
	"encoding/json"
	"time"

	"github.com/google/uuid"
)

type User struct {
	ID           uuid.UUID  `json:"id"`
	Username     string     `json:"username"`
	FullName     string     `json:"full_name"`
	DisplayName  *string    `json:"display_name"`
	Email        *string    `json:"email"`
	PasswordHash *string    `json:"password_hash"`
	Role         string     `json:"role"`
	Language     string     `json:"language"`
	AccountType  string     `json:"account_type"`
	TotpSecret   *string    `json:"totp_secret"`
	TotpEnabled  bool       `json:"totp_enabled"`
	IsActive     bool       `json:"is_active"`
	TimeFormat   string     `json:"time_format"`
	CreatedAt    time.Time  `json:"created_at"`
	UpdatedAt    time.Time  `json:"updated_at"`
}

type Session struct {
	ID        uuid.UUID  `json:"id"`
	UserID    uuid.UUID  `json:"user_id"`
	TokenHash string     `json:"token_hash"`
	IpAddress *string    `json:"ip_address"`
	UserAgent *string    `json:"user_agent"`
	ExpiresAt time.Time  `json:"expires_at"`
	CreatedAt time.Time  `json:"created_at"`
}

type OauthProvider struct {
	ID           uuid.UUID `json:"id"`
	Name         string    `json:"name"`
	ClientID     string    `json:"client_id"`
	ClientSecret string    `json:"client_secret"`
	AuthorizeUrl string    `json:"authorize_url"`
	TokenUrl     string    `json:"token_url"`
	UserinfoUrl  string    `json:"userinfo_url"`
	Scopes       string    `json:"scopes"`
	IsEnabled    bool      `json:"is_enabled"`
	CreatedAt    time.Time `json:"created_at"`
}

type OauthConnection struct {
	ID           uuid.UUID `json:"id"`
	UserID       uuid.UUID `json:"user_id"`
	ProviderID   uuid.UUID `json:"provider_id"`
	ExternalID   string    `json:"external_id"`
	AccessToken  *string   `json:"access_token"`
	RefreshToken *string   `json:"refresh_token"`
	CreatedAt    time.Time `json:"created_at"`
}

type RecoveryCode struct {
	ID       uuid.UUID  `json:"id"`
	UserID   uuid.UUID  `json:"user_id"`
	CodeHash string     `json:"code_hash"`
	UsedAt   *time.Time `json:"used_at"`
}

type Team struct {
	ID           uuid.UUID `json:"id"`
	Name         string    `json:"name"`
	Abbreviation string    `json:"abbreviation"`
	Color        string    `json:"color"`
	SortOrder    int32     `json:"sort_order"`
	IsActive     bool      `json:"is_active"`
	CreatedAt    time.Time `json:"created_at"`
}

type Event struct {
	ID               uuid.UUID  `json:"id"`
	Name             string     `json:"name"`
	Slug             string     `json:"slug"`
	Description      *string    `json:"description"`
	Location         *string    `json:"location"`
	ParticipantCount *int32     `json:"participant_count"`
	StartTime        time.Time  `json:"start_time"`
	EndTime          time.Time  `json:"end_time"`
	TimeGranularity  string     `json:"time_granularity"`
	IsLocked         bool       `json:"is_locked"`
	IsPublic         bool       `json:"is_public"`
	CreatedBy        *uuid.UUID `json:"created_by"`
	CreatedAt        time.Time  `json:"created_at"`
	UpdatedAt        time.Time  `json:"updated_at"`
}

type EventTeam struct {
	EventID   uuid.UUID `json:"event_id"`
	TeamID    uuid.UUID `json:"team_id"`
	IsVisible bool      `json:"is_visible"`
}

type EventAdmin struct {
	EventID uuid.UUID `json:"event_id"`
	UserID  uuid.UUID `json:"user_id"`
}

type EventHiddenRange struct {
	ID            uuid.UUID `json:"id"`
	EventID       uuid.UUID `json:"event_id"`
	HideStartHour int32     `json:"hide_start_hour"`
	HideEndHour   int32     `json:"hide_end_hour"`
}

type Shift struct {
	ID        uuid.UUID  `json:"id"`
	EventID   uuid.UUID  `json:"event_id"`
	TeamID    uuid.UUID  `json:"team_id"`
	UserID    uuid.UUID  `json:"user_id"`
	StartTime time.Time  `json:"start_time"`
	EndTime   time.Time  `json:"end_time"`
	CreatedBy *uuid.UUID `json:"created_by"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt time.Time  `json:"updated_at"`
}

type CoverageRequirement struct {
	ID            uuid.UUID `json:"id"`
	EventID       uuid.UUID `json:"event_id"`
	TeamID        uuid.UUID `json:"team_id"`
	StartTime     time.Time `json:"start_time"`
	EndTime       time.Time `json:"end_time"`
	RequiredCount int32     `json:"required_count"`
}

type UserAvailability struct {
	ID        uuid.UUID `json:"id"`
	EventID   uuid.UUID `json:"event_id"`
	UserID    uuid.UUID `json:"user_id"`
	StartTime time.Time `json:"start_time"`
	EndTime   time.Time `json:"end_time"`
	Status    string    `json:"status"`
	Note      *string   `json:"note"`
}

type IcalToken struct {
	ID         uuid.UUID  `json:"id"`
	UserID     uuid.UUID  `json:"user_id"`
	TokenHash  string     `json:"token_hash"`
	Token      string     `json:"token"`
	Label      string     `json:"label"`
	Scope      string     `json:"scope"`
	EventID    *uuid.UUID `json:"event_id"`
	TeamID     *uuid.UUID `json:"team_id"`
	CreatedAt  time.Time  `json:"created_at"`
	LastUsedAt *time.Time `json:"last_used_at"`
	IsActive   bool       `json:"is_active"`
}

type Notification struct {
	ID          uuid.UUID  `json:"id"`
	UserID      uuid.UUID  `json:"user_id"`
	EventID     *uuid.UUID `json:"event_id"`
	Title       string     `json:"title"`
	Body        *string    `json:"body"`
	TriggerType string     `json:"trigger_type"`
	IsRead      bool       `json:"is_read"`
	CreatedAt   time.Time  `json:"created_at"`
}

type NotificationPreference struct {
	UserID      uuid.UUID `json:"user_id"`
	TriggerType string    `json:"trigger_type"`
	Channel     string    `json:"channel"`
	IsEnabled   bool      `json:"is_enabled"`
}

type WebhookConfig struct {
	ID           uuid.UUID `json:"id"`
	EventID      uuid.UUID `json:"event_id"`
	Name         string    `json:"name"`
	Url          string    `json:"url"`
	Secret       string    `json:"secret"`
	TriggerTypes []string  `json:"trigger_types"`
	IsEnabled    bool      `json:"is_enabled"`
	CreatedAt    time.Time `json:"created_at"`
}

type AuditLog struct {
	ID         uuid.UUID       `json:"id"`
	UserID     *uuid.UUID      `json:"user_id"`
	EventID    *uuid.UUID      `json:"event_id"`
	Action     string          `json:"action"`
	EntityType string          `json:"entity_type"`
	EntityID   *uuid.UUID      `json:"entity_id"`
	OldValue   json.RawMessage `json:"old_value"`
	NewValue   json.RawMessage `json:"new_value"`
	IpAddress  *string         `json:"ip_address"`
	CreatedAt  time.Time       `json:"created_at"`
}

type SmtpConfig struct {
	ID          uuid.UUID `json:"id"`
	Host        string    `json:"host"`
	Port        int32     `json:"port"`
	Username    *string   `json:"username"`
	Password    *string   `json:"password"`
	FromAddress string    `json:"from_address"`
	FromName    *string   `json:"from_name"`
	UseTls      bool      `json:"use_tls"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type AppSetting struct {
	Key       string          `json:"key"`
	Value     json.RawMessage `json:"value"`
	UpdatedAt time.Time       `json:"updated_at"`
}