# api/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Rncasp API
  description: >
    REST API for the Rncasp (Really No Clue About Shift Planning) application.
    A shift planning web application for LAN party events that replaces Excel-based
    workflows where team members sign up for shifts across multi-day events.


    All responses (except health and binary downloads) are wrapped in a standard
    envelope: `{"data": ..., "error": ..., "meta": ...}`.


    Authentication is cookie-based using session cookies set on login/register.
  version: 1.0.0
  contact:
    name: Rncasp
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - cookieAuth: []
paths:
...
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Auth]
      operationId: updateProfile
      summary: Update own profile
      description: >
        Update the authenticated user's profile fields. All fields are
        optional; only provided fields are updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, user, read_only]
        - name: account_type
          in: query
          schema:
            type: string
            enum: [local, oauth, dummy]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/events/{slug}/export/csv:
    get:
      tags: [Export]
      operationId: exportCSV
      summary: Export event shifts as CSV
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/export/ical:
    get:
      tags: [Export]
      operationId: exportICal
      summary: Export event shifts as iCal
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: iCal file download
          content:
            text/calendar:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/export/pdf:
    get:
      tags: [Export]
      operationId: exportPDF
      summary: Export event shifts as PDF
      parameters:
        - $ref: "#/components/parameters/EventSlug"
        - name: layout
          in: query
          schema:
            type: string
            enum: [grid, list]
            default: grid
        - name: paper
          in: query
          schema:
            type: string
            enum: [A4, A3]
            default: A4
        - name: landscape
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: coverage
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: colors
          in: query
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: days
          in: query
          description: Comma-separated ISO dates (YYYY-MM-DD). Empty = all days.
          schema:
            type: string
        - name: users
          in: query
          description: Comma-separated user UUIDs. Empty = all users.
          schema:
            type: string
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Webhooks
  # ---------------------------------------------------------------------------
...
  /ical/{tokenId}/{token}:
    get:
      tags: [iCal]
      operationId: serveICalFeed
      summary: Serve iCal subscription feed (legacy)
      description: Public endpoint authenticated by token in the URL path.
      security: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: iCal feed
          content:
            text/calendar:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by the login or register endpoints.
        The cookie name is configurable; `session` is the default.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    EventSlug:
      name: slug
      in: path
      required: true
      description: URL-friendly event identifier
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
        minimum: 1
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
        minimum: 0

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: invalid_input
              message: "invalid request body"
    Unauthorized:
      description: Not authenticated or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: "not authenticated"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: "resource not found"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - not_found
            - already_exists
            - conflict
            - unauthorized
            - forbidden
            - invalid_input
            - event_locked
            - overbooking_not_allowed
            - account_inactive
            - dummy_account_cannot_login
            - invalid_totp_code
            - totp_required
            - too_many_requests
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Field name for validation errors
...
    User:
      type: object
      required:
        - id
        - username
        - full_name
        - role
        - language
        - account_type
        - time_format
        - totp_enabled
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        full_name:
          type: string
        display_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        role:
          type: string
          enum: [super_admin, user, read_only]
        language:
          type: string
          enum: [en, de]
        account_type:
          type: string
          enum: [local, oauth, dummy]
        time_format:
          type: string
          enum: ["24h", "12h"]
        totp_enabled:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"

    # --- Auth requests ---
...
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
        display_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
        time_format:
          type: string
          enum: ["24h", "12h"]

    # --- TOTP ---
...